requirements.txt
Pillow==11.0.0
reportlab==4.2.5
svglib==1.5.1
numpy==2.1.3
# fpdf==1.7.2
opencv-contrib-python==4.10.0.84
